+++ src/PhpWord/Writer/Word2007/Part/Settings.php
@@ -25,7 +25,7 @@ class Settings extends AbstractPart
     /**
      * Write word/settings.xml
      */
-    public function writeSettings()
+    public function write()
     {
         $settings = array(
             'w:zoom' => array('@attributes' => array('w:percent' => '100')),
@@ -135,5 +135,16 @@ class Settings extends AbstractPart
             }
             $xmlWriter->endElement();
         }
+    }
+
+    /**
+     * Write word/settings.xml
+     *
+     * @deprecated 0.11.0
+     * @codeCoverageIgnore
+     */
+    public function writeSettings()
+    {
+        return $this->write();
     }
 }